integrated the scripts
