+++ calm-integrations/ey-vm-migration-dr-scripts/ey-recreate-user-categories.py
@@ -32,13 +32,17 @@ logging.basicConfig(level=logging.INFO,
                     format="%(message)s",
                     datefmt='%H:%M:%S')
 
+SOURCE_PC_IP = "10.44.19.70"
 DESTINATION_PC_IP = "10.46.7.50"
 PC_PORT = 9440
 LENGTH = 100
-user_categories_list = []
 
+source_base_url = "https://{}:{}/api/nutanix/v3".format(SOURCE_PC_IP,str(PC_PORT))
 dest_base_url = "https://{}:{}/api/nutanix/v3".format(DESTINATION_PC_IP,str(PC_PORT))
+source_pc_auth = { "username": os.environ['SOURCE_PC_USER'], "password": os.environ['SOURCE_PC_PASS']}
 dest_pc_auth = { "username": os.environ['DEST_PC_USER'], "password": os.environ['DEST_PC_PASS']}
+
+user_categories_list = []
 
 CATEGORY_KEY_LIST = [
     "CalmApplication",
@@ -82,6 +86,31 @@ def create_category(base_url, auth, key):
         print('Response: {}'.format(json.dumps(json.loads(resp.content), indent=4)))
         exit(1)
 
+def get_category_values(base_url, auth, key, offset):
+    method = 'POST'
+    url = base_url + "/categories/{}/list".format(key)
+    category_value_list = []
+ 
+    payload = {"length": LENGTH, "offset": offset}
+    resp = requests.request(
+            method,
+            url,
+            data=json.dumps(payload),
+            headers=headers,
+            auth=(auth["username"], auth["password"]),
+            verify=False
+    )
+    if resp.ok:
+        resp_json = resp.json()
+        for entity in resp_json["entities"]:
+            category_value_list.append(entity["value"])
+        return resp_json["metadata"]["total_matches"], category_value_list
+    else:
+        print("Request to get category list for key '{}'.".format(key))
+        print('Status code: {}'.format(resp.status_code))
+        print('Response: {}'.format(json.dumps(json.loads(resp.content), indent=4)))
+        exit(1)
+
 def create_category_value(base_url, auth, key, value):
     method = 'PUT'
     url = base_url + "/categories/{}/{}".format(key, value)
@@ -106,6 +135,16 @@ def create_category_value(base_url, auth, key, value):
         exit(1)
 
 def main():
+    for key in CATEGORY_KEY_LIST:
+        offset = 0
+        while True:
+            matches, category_value_list = get_category_values(source_base_url, source_pc_auth, key, offset)
+            for value in category_value_list:
+                #print("key: {} - value: {}".format(key,value))
+                create_category_value(dest_base_url, dest_pc_auth, key, value)
+            offset += LENGTH
+            if (offset > matches):
+                break
     offset=0
     try:
         init_contexts()
